// src/components/logistics/types.ts
export interface InventoryItem {
  id: number;
  name: string;
  nombre_item?: string;
  category: string;
  size: string;
  quantity: number;
  min_stock: number;
  max_stock: number;
  purchase_price: number;
  sale_price: number;
  supplier: string;
  center: 'central' | 'sevilla' | 'jerez' | 'puerto';
  location: string;
  last_updated: string;
  status: 'in_stock' | 'low_stock' | 'out_of_stock';
}